==> data_preprocessing/Step 2 - Determining Neonates with all requirements .ipynb <==
import pandas as pd
from datetime import datetime
import seaborn as sns

import warnings
warnings.filterwarnings('ignore')

def side_by_side(*objs, **kwds):
    from pandas.io.formats.printing import adjoin
    space = kwds.get('space', 4)
    reprs = [repr(obj).split('\n') for obj in objs]
    print (adjoin(space, *reprs))
    print()
    return

## Load Data 

## 1.1 Heart Rate 

heart_rate_df = pd.read_csv("Data_3173_TS/heart_rate.csv")  

heart_rate_df.head()

heart_rate_df['HADM_ID'].nunique()

heart_rate_df.info()

start = datetime.now()
heart_rate_df['CHARTTIME'] = heart_rate_df['CHARTTIME'].astype('datetime64[ns]')
heart_rate_df['ADMITTIME'] = heart_rate_df['ADMITTIME'].astype('datetime64[ns]')

print("Computing time takes %s" %(datetime.now()-start))

## Sort dataframe 
heart_rate_df.sort_values(by=['HADM_ID', 'CHARTTIME'], inplace=True)  

## Giving row number to every row 
heart_rate_df['row'] = heart_rate_df.groupby(['HADM_ID']).cumcount()+1

## Take the first row 
heart_rate_1 = heart_rate_df[heart_rate_df.row == 1]

heart_rate_1 = heart_rate_df[heart_rate_df.row == 1]

heart_rate_1.head()

heart_rate_1.columns 

heart_rate_1['DIFF'] = heart_rate_1['CHARTTIME'] - heart_rate_1['ADMITTIME']

heart_rate_1.describe()

heart_rate_1['TIME_INTERVAL'] = heart_rate_1['CHARTTIME'] - heart_rate_1['ADMITTIME']

heart_rate_1[(heart_rate_1.TIME_INTERVAL <= '1 days') & (heart_rate_1.TIME_INTERVAL >= '-1 days')]

heart_rate_1.shape 

admissions = heart_rate_1['HADM_ID']
admissions = admissions.to_frame()

type(admissions)

heart_rate_24 = heart_rate_1[(heart_rate_1.TIME_INTERVAL <= '1 days') & (heart_rate_1.TIME_INTERVAL >= '-1 days')]

heart_rate_24['HADM_ID'].nunique()

admissions['HeartRate'] = admissions['HADM_ID'].isin(heart_rate_24['HADM_ID'])
admissions.head()

admissions[admissions['HeartRate'] == True].shape

## 1.2 Resp Rate 

resp_rate_df = pd.read_csv("Data_3173_TS/resp_rate.csv")  

resp_rate_df['CHARTTIME'] = resp_rate_df['CHARTTIME'].astype('datetime64[ns]')
resp_rate_df['ADMITTIME'] = resp_rate_df['ADMITTIME'].astype('datetime64[ns]')

## Sort dataframe 
resp_rate_df.sort_values(by=['HADM_ID', 'CHARTTIME'], inplace=True)  

## Giving row number to every row 
resp_rate_df['row'] = resp_rate_df.groupby(['HADM_ID']).cumcount()+1

## Take the first row 
resp_rate_1 = resp_rate_df[resp_rate_df.row == 1]

resp_rate_1['TIME_INTERVAL'] = resp_rate_1['CHARTTIME'] - resp_rate_1['ADMITTIME'] 

resp_rate_24 = resp_rate_1[(resp_rate_1.TIME_INTERVAL <= '1 days') & (resp_rate_1.TIME_INTERVAL >= '-1 days')]

resp_rate_24.shape

admissions['RespRate'] = admissions['HADM_ID'].isin(resp_rate_24['HADM_ID'])
admissions.head()

## 1.3 SaO2 

sao2_df = pd.read_csv("Data_3173_TS/sao2.csv")  

sao2_df['CHARTTIME'] = sao2_df['CHARTTIME'].astype('datetime64[ns]')
sao2_df['ADMITTIME'] = sao2_df['ADMITTIME'].astype('datetime64[ns]')

## Sort dataframe 
sao2_df.sort_values(by=['HADM_ID', 'CHARTTIME'], inplace=True)  

## Giving row number to every row 
sao2_df['row'] = sao2_df.groupby(['HADM_ID']).cumcount()+1

## Take the first row 
sao2_1 = sao2_df[sao2_df.row == 1]

sao2_1['TIME_INTERVAL'] = sao2_1['CHARTTIME'] - sao2_1['ADMITTIME'] 

sao2_24 = sao2_1[(sao2_1.TIME_INTERVAL <= '1 days') & (sao2_1.TIME_INTERVAL >= '-1 days')]

sao2_24.shape

admissions['SaO2'] = admissions['HADM_ID'].isin(sao2_24['HADM_ID'])
admissions.head()

## 1.4 HR Alarm High

hr_alarm_high_df = pd.read_csv("Data_3173_TS/hr_alarm_high.csv")  

hr_alarm_high_df['CHARTTIME'] = hr_alarm_high_df['CHARTTIME'].astype('datetime64[ns]')
hr_alarm_high_df['ADMITTIME'] = hr_alarm_high_df['ADMITTIME'].astype('datetime64[ns]')

## Sort dataframe 
hr_alarm_high_df.sort_values(by=['HADM_ID', 'CHARTTIME'], inplace=True)  

## Giving row number to every row 
hr_alarm_high_df['row'] = hr_alarm_high_df.groupby(['HADM_ID']).cumcount()+1

## Take the first row 
hr_alarm_high_1 = hr_alarm_high_df[hr_alarm_high_df.row == 1]

hr_alarm_high_1['TIME_INTERVAL'] = hr_alarm_high_1['CHARTTIME'] - hr_alarm_high_1['ADMITTIME'] 

hr_alarm_high_24 = hr_alarm_high_1[(hr_alarm_high_1.TIME_INTERVAL <= '1 days') & (hr_alarm_high_1.TIME_INTERVAL >= '-1 days')]

hr_alarm_high_24.shape

admissions['hrAlarmHigh'] = admissions['HADM_ID'].isin(hr_alarm_high_24['HADM_ID'])
admissions.head()

## 1.5 HR Alarm Low 

hr_alarm_low_df = pd.read_csv("Data_3173_TS/hr_alarm_low.csv")  

hr_alarm_low_df['CHARTTIME'] = hr_alarm_low_df['CHARTTIME'].astype('datetime64[ns]')
hr_alarm_low_df['ADMITTIME'] = hr_alarm_low_df['ADMITTIME'].astype('datetime64[ns]')

## Sort dataframe 
hr_alarm_low_df.sort_values(by=['HADM_ID', 'CHARTTIME'], inplace=True)  

## Giving row number to every row 
hr_alarm_low_df['row'] = hr_alarm_low_df.groupby(['HADM_ID']).cumcount()+1

## Take the first row 
hr_alarm_low_1 = hr_alarm_low_df[hr_alarm_low_df.row == 1]

hr_alarm_low_1['TIME_INTERVAL'] = hr_alarm_low_1['CHARTTIME'] - hr_alarm_low_1['ADMITTIME'] 

hr_alarm_low_24 = hr_alarm_low_1[(hr_alarm_low_1.TIME_INTERVAL <= '1 days') & (hr_alarm_low_1.TIME_INTERVAL >= '-1 days')]

hr_alarm_low_24.shape

admissions['hrAlarmLow'] = admissions['HADM_ID'].isin(hr_alarm_low_24['HADM_ID'])
admissions.head()

## 1.6 SaO2 Alarm High 

sao2_alarm_high_df = pd.read_csv("Data_3173_TS/sao2_alarm_high.csv")  

sao2_alarm_high_df['CHARTTIME'] = sao2_alarm_high_df['CHARTTIME'].astype('datetime64[ns]')
sao2_alarm_high_df['ADMITTIME'] = sao2_alarm_high_df['ADMITTIME'].astype('datetime64[ns]')

## Sort dataframe 
sao2_alarm_high_df.sort_values(by=['HADM_ID', 'CHARTTIME'], inplace=True)  

## Giving row number to every row 
sao2_alarm_high_df['row'] = sao2_alarm_high_df.groupby(['HADM_ID']).cumcount()+1

## Take the first row 
sao2_alarm_high_1 = sao2_alarm_high_df[sao2_alarm_high_df.row == 1]

sao2_alarm_high_1['TIME_INTERVAL'] = sao2_alarm_high_1['CHARTTIME'] - sao2_alarm_high_1['ADMITTIME'] 

sao2_alarm_high_24 = sao2_alarm_high_1[(sao2_alarm_high_1.TIME_INTERVAL <= '1 days') & (sao2_alarm_high_1.TIME_INTERVAL >= '-1 days')]

sao2_alarm_high_24.shape

admissions['sao2AlarmHigh'] = admissions['HADM_ID'].isin(sao2_alarm_high_24['HADM_ID'])
admissions.head()

## 1.7 SaO2 Alarm Low 

sao2_alarm_low_df = pd.read_csv("Data_3173_TS/sao2_alarm_low.csv")  

sao2_alarm_low_df['CHARTTIME'] = sao2_alarm_low_df['CHARTTIME'].astype('datetime64[ns]')
sao2_alarm_low_df['ADMITTIME'] = sao2_alarm_low_df['ADMITTIME'].astype('datetime64[ns]')

## Sort dataframe 
sao2_alarm_low_df.sort_values(by=['HADM_ID', 'CHARTTIME'], inplace=True)  

## Giving row number to every row 
sao2_alarm_low_df['row'] = sao2_alarm_low_df.groupby(['HADM_ID']).cumcount()+1

## Take the first row 
sao2_alarm_low_1 = sao2_alarm_low_df[sao2_alarm_low_df.row == 1]

sao2_alarm_low_1['TIME_INTERVAL'] = sao2_alarm_low_1['CHARTTIME'] - sao2_alarm_low_1['ADMITTIME'] 

sao2_alarm_low_24 = sao2_alarm_low_1[(sao2_alarm_low_1.TIME_INTERVAL <= '1 days') & (sao2_alarm_low_1.TIME_INTERVAL >= '-1 days')]

sao2_alarm_low_24.shape

admissions['sao2AlarmLow'] = admissions['HADM_ID'].isin(sao2_alarm_low_24['HADM_ID'])
admissions.head()

## 1.8 Temperature 

temperature_df = pd.read_csv("Data_3173_TS/temperature.csv")  

temperature_df['CHARTTIME'] = temperature_df['CHARTTIME'].astype('datetime64[ns]')
temperature_df['ADMITTIME'] = temperature_df['ADMITTIME'].astype('datetime64[ns]')

## Sort dataframe 
temperature_df.sort_values(by=['HADM_ID', 'CHARTTIME'], inplace=True)  

## Giving row number to every row 
temperature_df['row'] = temperature_df.groupby(['HADM_ID']).cumcount()+1

## Take the first row 
temperature_1 = temperature_df[temperature_df.row == 1]

temperature_1['TIME_INTERVAL'] = temperature_1['CHARTTIME'] - temperature_1['ADMITTIME'] 

temperature_24 = temperature_1[(temperature_1.TIME_INTERVAL <= '1 days') & (temperature_1.TIME_INTERVAL >= '-1 days')]

temperature_24.shape

admissions['temperature'] = admissions['HADM_ID'].isin(temperature_24['HADM_ID'])
admissions.head()

## 1.9 Skin Temperature 

skin_temperature_df = pd.read_csv("Data_3173_TS/skin_temperature.csv")  

skin_temperature_df['CHARTTIME'] = skin_temperature_df['CHARTTIME'].astype('datetime64[ns]')
skin_temperature_df['ADMITTIME'] = skin_temperature_df['ADMITTIME'].astype('datetime64[ns]')

## Sort dataframe 
skin_temperature_df.sort_values(by=['HADM_ID', 'CHARTTIME'], inplace=True)  

## Giving row number to every row 
skin_temperature_df['row'] = skin_temperature_df.groupby(['HADM_ID']).cumcount()+1

## Take the first row 
skin_temperature_1 = skin_temperature_df[skin_temperature_df.row == 1]

skin_temperature_1['TIME_INTERVAL'] = skin_temperature_1['CHARTTIME'] - skin_temperature_1['ADMITTIME'] 

skin_temperature_24 = skin_temperature_1[(skin_temperature_1.TIME_INTERVAL <= '1 days') & (skin_temperature_1.TIME_INTERVAL >= '-1 days')]

skin_temperature_24.shape

admissions['SkinTemperature'] = admissions['HADM_ID'].isin(skin_temperature_24['HADM_ID'])
admissions.head()

## 1.10 BP Cuff Diastolic 

bp_cuff_diastolic_df = pd.read_csv("Data_3173_TS/bp_cuff_diastolic.csv")  

bp_cuff_diastolic_df['CHARTTIME'] = bp_cuff_diastolic_df['CHARTTIME'].astype('datetime64[ns]')
bp_cuff_diastolic_df['ADMITTIME'] = bp_cuff_diastolic_df['ADMITTIME'].astype('datetime64[ns]')

## Sort dataframe 
bp_cuff_diastolic_df.sort_values(by=['HADM_ID', 'CHARTTIME'], inplace=True)  

## Giving row number to every row 
bp_cuff_diastolic_df['row'] = bp_cuff_diastolic_df.groupby(['HADM_ID']).cumcount()+1

## Take the first row 
bp_cuff_diastolic_1 = bp_cuff_diastolic_df[bp_cuff_diastolic_df.row == 1]

bp_cuff_diastolic_1['TIME_INTERVAL'] = bp_cuff_diastolic_1['CHARTTIME'] - bp_cuff_diastolic_1['ADMITTIME'] 

bp_cuff_diastolic_24 = bp_cuff_diastolic_1[(bp_cuff_diastolic_1.TIME_INTERVAL <= '1 days') & (bp_cuff_diastolic_1.TIME_INTERVAL >= '-1 days')]

bp_cuff_diastolic_24.shape

admissions['BPCuffDiastolic'] = admissions['HADM_ID'].isin(bp_cuff_diastolic_24['HADM_ID'])
admissions.head()

## 1.11 BP Cuff Mean 

bp_cuff_mean_df = pd.read_csv("Data_3173_TS/bp_cuff_mean.csv")  

bp_cuff_mean_df['CHARTTIME'] = bp_cuff_mean_df['CHARTTIME'].astype('datetime64[ns]')
bp_cuff_mean_df['ADMITTIME'] = bp_cuff_mean_df['ADMITTIME'].astype('datetime64[ns]')

## Sort dataframe 
bp_cuff_mean_df.sort_values(by=['HADM_ID', 'CHARTTIME'], inplace=True)  

## Giving row number to every row 
bp_cuff_mean_df['row'] = bp_cuff_mean_df.groupby(['HADM_ID']).cumcount()+1

## Take the first row 
bp_cuff_mean_1 = bp_cuff_mean_df[bp_cuff_mean_df.row == 1]

bp_cuff_mean_1['TIME_INTERVAL'] = bp_cuff_mean_1['CHARTTIME'] - bp_cuff_mean_1['ADMITTIME'] 

bp_cuff_mean_24 = bp_cuff_mean_1[(bp_cuff_mean_1.TIME_INTERVAL <= '1 days') & (bp_cuff_mean_1.TIME_INTERVAL >= '-1 days')]

bp_cuff_mean_24.shape

admissions['BPCuffMean'] = admissions['HADM_ID'].isin(bp_cuff_mean_24['HADM_ID'])
admissions.head()

## 1.12 BP Cuff Systolic 

bp_cuff_systolic_df = pd.read_csv("Data_3173_TS/bp_cuff_systolic.csv")  

bp_cuff_systolic_df['CHARTTIME'] = bp_cuff_systolic_df['CHARTTIME'].astype('datetime64[ns]')
bp_cuff_systolic_df['ADMITTIME'] = bp_cuff_systolic_df['ADMITTIME'].astype('datetime64[ns]')

## Sort dataframe 
bp_cuff_systolic_df.sort_values(by=['HADM_ID', 'CHARTTIME'], inplace=True)  

## Giving row number to every row 
bp_cuff_systolic_df['row'] = bp_cuff_systolic_df.groupby(['HADM_ID']).cumcount()+1

## Take the first row 
bp_cuff_systolic_1 = bp_cuff_systolic_df[bp_cuff_systolic_df.row == 1]

bp_cuff_systolic_1['TIME_INTERVAL'] = bp_cuff_systolic_1['CHARTTIME'] - bp_cuff_systolic_1['ADMITTIME'] 

bp_cuff_systolic_24 = bp_cuff_systolic_1[(bp_cuff_systolic_1.TIME_INTERVAL <= '1 days') & (bp_cuff_systolic_1.TIME_INTERVAL >= '-1 days')]

bp_cuff_systolic_24.shape

admissions['BPCuffSystolic'] = admissions['HADM_ID'].isin(bp_cuff_systolic_24['HADM_ID'])
admissions.head()

## 1.13 Glucometer

glucometer_df = pd.read_csv("Data_3173_TS/glucometer.csv")  

glucometer_df['CHARTTIME'] = glucometer_df['CHARTTIME'].astype('datetime64[ns]')
glucometer_df['ADMITTIME'] = glucometer_df['ADMITTIME'].astype('datetime64[ns]')

## Sort dataframe 
glucometer_df.sort_values(by=['HADM_ID', 'CHARTTIME'], inplace=True)  

## Giving row number to every row 
glucometer_df['row'] = glucometer_df.groupby(['HADM_ID']).cumcount()+1

## Take the first row 
glucometer_1 = glucometer_df[glucometer_df.row == 1]

glucometer_1['TIME_INTERVAL'] = glucometer_1['CHARTTIME'] - glucometer_1['ADMITTIME'] 

glucometer_24 = glucometer_1[(glucometer_1.TIME_INTERVAL <= '1 days') & (glucometer_1.TIME_INTERVAL >= '-1 days')]

glucometer_24.shape

admissions['Glucometer'] = admissions['HADM_ID'].isin(glucometer_24['HADM_ID'])
admissions.head()

# Other Chart Events 

## 2.1 Birth Weight 

birth_weight_df = pd.read_csv("Data_3173_SV/birth_weight.csv")  

birth_weight_df['CHARTTIME'] = birth_weight_df['CHARTTIME'].astype('datetime64[ns]')
birth_weight_df['ADMITTIME'] = birth_weight_df['ADMITTIME'].astype('datetime64[ns]')

## Sort dataframe 
birth_weight_df.sort_values(by=['HADM_ID', 'CHARTTIME'], inplace=True)  

## Giving row number to every row 
birth_weight_df['row'] = birth_weight_df.groupby(['HADM_ID']).cumcount()+1

## Take the first row 
birth_weight_1 = birth_weight_df[birth_weight_df.row == 1]

birth_weight_1['TIME_INTERVAL'] = birth_weight_1['CHARTTIME'] - birth_weight_1['ADMITTIME'] 

birth_weight_24 = birth_weight_1[(birth_weight_1.TIME_INTERVAL <= '1 days') & (birth_weight_1.TIME_INTERVAL >= '-1 days')]

birth_weight_24.shape

admissions['BirthWeight'] = admissions['HADM_ID'].isin(birth_weight_1['HADM_ID'])
admissions.head()

## 2.2 Head Circ 

head_circ_df = pd.read_csv("Data_3173_SV/head_circ.csv")  

head_circ_df['CHARTTIME'] = head_circ_df['CHARTTIME'].astype('datetime64[ns]')
head_circ_df['ADMITTIME'] = head_circ_df['ADMITTIME'].astype('datetime64[ns]')

## Sort dataframe 
head_circ_df.sort_values(by=['HADM_ID', 'CHARTTIME'], inplace=True)  

## Giving row number to every row 
head_circ_df['row'] = head_circ_df.groupby(['HADM_ID']).cumcount()+1

## Take the first row 
head_circ_1 = head_circ_df[head_circ_df.row == 1]

head_circ_1['TIME_INTERVAL'] = head_circ_1['CHARTTIME'] - head_circ_1['ADMITTIME'] 

head_circ_24 = head_circ_1[(head_circ_1.TIME_INTERVAL <= '10 days') & (head_circ_1.TIME_INTERVAL >= '-1 days')]

head_circ_24.shape

admissions['HeadCirc'] = admissions['HADM_ID'].isin(head_circ_1['HADM_ID'])
admissions.shape 

# Blood Test 

## 3.1 Bands 

bands_df = pd.read_csv("Data_3173_SV/bands.csv")  

bands_df['CHARTTIME'] = bands_df['CHARTTIME'].astype('datetime64[ns]')
bands_df['ADMITTIME'] = bands_df['ADMITTIME'].astype('datetime64[ns]')

## Sort dataframe 
bands_df.sort_values(by=['HADM_ID', 'CHARTTIME'], inplace=True)  

## Giving row number to every row 
bands_df['row'] = bands_df.groupby(['HADM_ID']).cumcount()+1

## Take the first row 
bands_1 = bands_df[bands_df.row == 1]

bands_1['TIME_INTERVAL'] = bands_1['CHARTTIME'] - bands_1['ADMITTIME'] 

bands_24 = bands_1[(bands_1.TIME_INTERVAL <= '1 days') & (bands_1.TIME_INTERVAL >= '-1 days')]

bands_24.shape

admissions['Bands'] = admissions['HADM_ID'].isin(bands_24['HADM_ID'])
admissions.shape 

## 3.2 Eosinophils

eosinophils_df = pd.read_csv("Data_3173_SV/eosinophils.csv")  

eosinophils_df['CHARTTIME'] = eosinophils_df['CHARTTIME'].astype('datetime64[ns]')
eosinophils_df['ADMITTIME'] = eosinophils_df['ADMITTIME'].astype('datetime64[ns]')

## Sort dataframe 
eosinophils_df.sort_values(by=['HADM_ID', 'CHARTTIME'], inplace=True)  

## Giving row number to every row 
eosinophils_df['row'] = eosinophils_df.groupby(['HADM_ID']).cumcount()+1

## Take the first row 
eosinophils_1 = eosinophils_df[eosinophils_df.row == 1]

eosinophils_1['TIME_INTERVAL'] = eosinophils_1['CHARTTIME'] - eosinophils_1['ADMITTIME'] 

eosinophils_24 = eosinophils_1[(eosinophils_1.TIME_INTERVAL <= '1 days') & (eosinophils_1.TIME_INTERVAL >= '-1 days')]

eosinophils_24.shape

admissions['Eosinophils'] = admissions['HADM_ID'].isin(eosinophils_24['HADM_ID'])
admissions.shape 

## 3.3 Lymphs

lymphs_df = pd.read_csv("Data_3173_SV/lymphs.csv")  

lymphs_df['CHARTTIME'] = lymphs_df['CHARTTIME'].astype('datetime64[ns]')
lymphs_df['ADMITTIME'] = lymphs_df['ADMITTIME'].astype('datetime64[ns]')

## Sort dataframe 
lymphs_df.sort_values(by=['HADM_ID', 'CHARTTIME'], inplace=True)  

## Giving row number to every row 
lymphs_df['row'] = lymphs_df.groupby(['HADM_ID']).cumcount()+1

## Take the first row 
lymphs_1 = lymphs_df[lymphs_df.row == 1]

lymphs_1['TIME_INTERVAL'] = lymphs_1['CHARTTIME'] - lymphs_1['ADMITTIME'] 

lymphs_24 = lymphs_1[(lymphs_1.TIME_INTERVAL <= '1 days') & (lymphs_1.TIME_INTERVAL >= '-1 days')]

lymphs_24.shape

admissions['Lymphs'] = admissions['HADM_ID'].isin(lymphs_24['HADM_ID'])
admissions.shape 

## 3.4 MONOs

monos_df = pd.read_csv("Data_3173_SV/monos.csv")  

monos_df['CHARTTIME'] = monos_df['CHARTTIME'].astype('datetime64[ns]')
monos_df['ADMITTIME'] = monos_df['ADMITTIME'].astype('datetime64[ns]')

## Sort dataframe 
monos_df.sort_values(by=['HADM_ID', 'CHARTTIME'], inplace=True)  

## Giving row number to every row 
monos_df['row'] = monos_df.groupby(['HADM_ID']).cumcount()+1

## Take the first row 
monos_1 = monos_df[monos_df.row == 1]

monos_1['TIME_INTERVAL'] = monos_1['CHARTTIME'] - monos_1['ADMITTIME'] 

monos_24 = monos_1[(monos_1.TIME_INTERVAL <= '1 days') & (monos_1.TIME_INTERVAL >= '-1 days')]

monos_24.shape

admissions['Monos'] = admissions['HADM_ID'].isin(monos_24['HADM_ID'])
admissions.shape 

## 3.5 NEUTs 

neuts_df = pd.read_csv("Data_3173_SV/neuts.csv")  

neuts_df['CHARTTIME'] = neuts_df['CHARTTIME'].astype('datetime64[ns]')
neuts_df['ADMITTIME'] = neuts_df['ADMITTIME'].astype('datetime64[ns]')

## Sort dataframe 
neuts_df.sort_values(by=['HADM_ID', 'CHARTTIME'], inplace=True)  

## Giving row number to every row 
neuts_df['row'] = neuts_df.groupby(['HADM_ID']).cumcount()+1

## Take the first row 
neuts_1 = neuts_df[neuts_df.row == 1]

neuts_1['TIME_INTERVAL'] = neuts_1['CHARTTIME'] - neuts_1['ADMITTIME'] 

neuts_24 = neuts_1[(neuts_1.TIME_INTERVAL <= '1 days') & (neuts_1.TIME_INTERVAL >= '-1 days')]

neuts_24.shape

admissions['Neuts'] = admissions['HADM_ID'].isin(neuts_24['HADM_ID'])
admissions.shape 

## 3.6 Platelet

platelet_df = pd.read_csv("Data_3173_SV/platelet.csv")  

platelet_df['CHARTTIME'] = platelet_df['CHARTTIME'].astype('datetime64[ns]')
platelet_df['ADMITTIME'] = platelet_df['ADMITTIME'].astype('datetime64[ns]')

## Sort dataframe 
platelet_df.sort_values(by=['HADM_ID', 'CHARTTIME'], inplace=True)  

## Giving row number to every row 
platelet_df['row'] = platelet_df.groupby(['HADM_ID']).cumcount()+1

## Take the first row 
platelet_1 = platelet_df[platelet_df.row == 1]

platelet_1['TIME_INTERVAL'] = platelet_1['CHARTTIME'] - platelet_1['ADMITTIME'] 

platelet_24 = platelet_1[(platelet_1.TIME_INTERVAL <= '1 days') & (platelet_1.TIME_INTERVAL >= '-1 days')]

platelet_24.shape

admissions['Platelet'] = admissions['HADM_ID'].isin(platelet_24['HADM_ID'])
admissions.shape 

admissions.to_csv("admissions_3K.csv", index=False)
