==> data_preprocessing/Step 3 - Extracting first 24 hour (with minus) .ipynb <==
import pandas as pd
from datetime import datetime
import seaborn as sns

import warnings
warnings.filterwarnings('ignore')

def side_by_side(*objs, **kwds):
    from pandas.io.formats.printing import adjoin
    space = kwds.get('space', 4)
    reprs = [repr(obj).split('\n') for obj in objs]
    print (adjoin(space, *reprs))
    print()
    return

## Load Data 

admissions = pd.read_csv("admissions_3K.csv")

admissions.shape 

admissions.columns

master_neonates = admissions.loc[(admissions.HeartRate==True) & (admissions.RespRate==True) & (admissions.SaO2==True) 
               & (admissions.hrAlarmHigh==True) & (admissions.hrAlarmLow==True) & (admissions.sao2AlarmHigh==True)
               & (admissions.sao2AlarmLow==True) & (admissions.BPCuffDiastolic==True) & (admissions.BPCuffMean==True)
               & (admissions.BPCuffSystolic==True) & (admissions.SkinTemperature==True) & (admissions.temperature==True)
               & (admissions.HeadCirc==True) & (admissions.Glucometer==True)
               & (admissions.Eosinophils==True)
               & (admissions.Lymphs==True) & (admissions.Monos==True) & (admissions.Neuts==True)
               & (admissions.Bands==True) & (admissions.Platelet==True)
               & (admissions.BirthWeight==True)
               ] 

master_neonates.shape 

master_neonates = master_neonates['HADM_ID'].to_frame()

## Defining Extracting Function 

def extract_24(df, master_ids, filename):
    ## Renaming columns 
    df.rename(columns = {'subject_id_x': 'subject_id'}, inplace=True) 
    
    ## Dropping extra columns 
    df.drop(['subject_id_y', 'DISCHTIME', 'Dead'], axis=1, inplace=True)
    
    ## Merging with master neonates 
    df = pd.merge(df, master_ids, on='HADM_ID', how='inner')
    
    ## Converting dates 
    df['CHARTTIME'] = df['CHARTTIME'].astype('datetime64[ns]')
    df['ADMITTIME'] = df['ADMITTIME'].astype('datetime64[ns]') 
    
    ## Calculating time interval 
    df['TIME_INTERVAL'] = df['CHARTTIME'] - df['ADMITTIME']

    ## Extracting only 24 hours data 
    df_24 = df[(df.TIME_INTERVAL <= '1 days') & (df.TIME_INTERVAL >= '-1 days')]

    df_24.drop(['TIME_INTERVAL'], axis=1, inplace=True) 
    
    ## Saving into a given file 
    df_24.to_csv(filename, index=False)
    

## 1 Load Chart Events 

## 1.1 Heart Rate 

heart_rate_df = pd.read_csv("Data_3173_TS/heart_rate.csv")  

heart_rate_df.head()

heart_rate_df.rename(columns = {'subject_id_x': 'subject_id'}, inplace=True) 
heart_rate_df.drop(['subject_id_y', 'DISCHTIME', 'Dead'], axis=1, inplace=True)

heart_rate_df.head()

heart_rate_df['HADM_ID'].nunique()

heart_rate_df.shape 

heart_rate_df = pd.merge(heart_rate_df, master_neonates, on='HADM_ID', how='inner')

heart_rate_df['HADM_ID'].nunique()

heart_rate_df.shape 

heart_rate_df.head(2)

start = datetime.now()
heart_rate_df['CHARTTIME'] = heart_rate_df['CHARTTIME'].astype('datetime64[ns]')
heart_rate_df['ADMITTIME'] = heart_rate_df['ADMITTIME'].astype('datetime64[ns]')

print("Computing time takes %s" %(datetime.now()-start))

heart_rate_df['TIME_INTERVAL'] = heart_rate_df['CHARTTIME'] - heart_rate_df['ADMITTIME']

heart_rate_24 = heart_rate_df[(heart_rate_df.TIME_INTERVAL <= '1 days') & (heart_rate_df.TIME_INTERVAL >= '-1 days')]

heart_rate_24.shape 

heart_rate_24['HADM_ID'].nunique()

heart_rate_24.to_csv('Data_3145_TS_24/heart_rate_24.csv', index=False)

extract_24(heart_rate_df, master_neonates, 'Data_3145_TS_24/heart_rate_24.csv')

## 1.2 Resp Rate 

resp_rate_df = pd.read_csv("Data_3173_TS/resp_rate.csv") 
resp_rate_df.head(2)

extract_24(resp_rate_df, master_neonates, 'Data_3145_TS_24/resp_rate_24.csv')

## 1.3 SaO2 

sao2_df = pd.read_csv("Data_3173_TS/sao2.csv")  
sao2_df.head(2)

extract_24(sao2_df, master_neonates, 'Data_3145_TS_24/sao2_24.csv')

## 1.4 HR Alarm High 

hr_alarm_high_df = pd.read_csv("Data_3173_TS/hr_alarm_high.csv")  
hr_alarm_high_df.head(2)

extract_24(hr_alarm_high_df, master_neonates, 'Data_3145_TS_24/hr_alarm_high_24.csv')

## 1.5 HR Alarm Low 

hr_alarm_low_df = pd.read_csv("Data_3173_TS/hr_alarm_low.csv")  
extract_24(hr_alarm_low_df, master_neonates, 'Data_3145_TS_24/hr_alarm_low_24.csv')

## 1.6 SaO2 Alarm High 

sao2_alarm_high_df = pd.read_csv("Data_3173_TS/sao2_alarm_high.csv")  
extract_24(sao2_alarm_high_df, master_neonates, 'Data_3145_TS_24/sao2_alarm_high_24.csv')

## 1.7 SaO2 Alarm Low 

sao2_alarm_low_df = pd.read_csv("Data_3173_TS/sao2_alarm_low.csv")  
extract_24(sao2_alarm_low_df, master_neonates, 'Data_3145_TS_24/sao2_alarm_low_24.csv')

## 1.8 BP Cuff Diastolic 

bp_cuff_diastolic_df = pd.read_csv("Data_3173_TS/bp_cuff_diastolic.csv")  
extract_24(bp_cuff_diastolic_df, master_neonates, 'Data_3145_TS_24/bp_cuff_diastolic_24.csv')

## 1.9 BP Cuff Mean 

bp_cuff_mean_df = pd.read_csv("Data_3173_TS/bp_cuff_mean.csv")  
extract_24(bp_cuff_mean_df, master_neonates, 'Data_3145_TS_24/bp_cuff_mean_24.csv')

## 1.10 BP Cuff Systolic 

bp_cuff_systolic_df = pd.read_csv("Data_3173_TS/bp_cuff_systolic.csv")  
extract_24(bp_cuff_systolic_df, master_neonates, 'Data_3145_TS_24/bp_cuff_systolic_24.csv')

## 1.11 Temperature 

temperature_df = pd.read_csv("Data_3173_TS/temperature.csv")  
extract_24(temperature_df, master_neonates, 'Data_3145_TS_24/temperature_24.csv')

## 1.12 Skin Temperature 

skin_temperature_df = pd.read_csv("Data_3173_TS/skin_temperature.csv")  
extract_24(skin_temperature_df, master_neonates, 'Data_3145_TS_24/skin_temperature_24.csv')

## 1.13 Glucometer

glucometer_df = pd.read_csv("Data_3173_TS/glucometer.csv")  
extract_24(glucometer_df, master_neonates, 'Data_3145_TS_24/glucometer_24.csv')

## Other Chart Event Features 

def extract_24_bw(df, master_ids, filename):
    ## Renaming columns 
    df.rename(columns = {'subject_id_x': 'subject_id'}, inplace=True) 
    
    ## Dropping extra columns 
    df.drop(['subject_id_y', 'DISCHTIME', 'Dead'], axis=1, inplace=True)
    
    ## Merging with master neonates 
    df = pd.merge(df, master_ids, on='HADM_ID', how='inner')
    
    ## Saving into a given file 
    df.to_csv(filename, index=False)

## 2.1 Birth Weight 

birth_weight_df = pd.read_csv("Data_3173_SV/birth_weight.csv") 

birth_weight_df.head(2)

extract_24_bw(birth_weight_df, master_neonates,'Data_3145_SV_24/birth_weight_24.csv')

## 2.2 Head Circ 

head_circ_df = pd.read_csv("Data_3173_SV/head_circ.csv") 

head_circ_df.head(2)

extract_24_bw(head_circ_df, master_neonates,'Data_3145_SV_24/head_circ_24.csv')

## Blood Test 

## 3.1 Bands 

bands_df = pd.read_csv("Data_3173_SV/bands.csv") 

bands_df.head(2)

extract_24(bands_df, master_neonates,'Data_3145_SV_24/bands_24.csv')

## 3.2 Eosinophils

eosinophils_df = pd.read_csv("Data_3173_SV/eosinophils.csv")  
eosinophils_df.head(2)

extract_24(eosinophils_df, master_neonates,'Data_3145_SV_24/eosinophils_24.csv')

## 3.3 Lymphs

lymphs_df = pd.read_csv("Data_3173_SV/lymphs.csv")  
extract_24(lymphs_df, master_neonates,'Data_3145_SV_24/lymphs_24.csv')

## 3.4 MONOs

monos_df = pd.read_csv("Data_3173_SV/monos.csv")  
extract_24(monos_df, master_neonates,'Data_3145_SV_24/monos_24.csv')

## 3.5 NEUTs

neuts_df = pd.read_csv("Data_3173_SV/neuts.csv")  
extract_24(neuts_df, master_neonates,'Data_3145_SV_24/neuts_24.csv')

## 3.6 Platelet

platelet_df = pd.read_csv("Data_3173_SV/platelet.csv")  
extract_24(platelet_df, master_neonates,'Data_3145_SV_24/platelet_24.csv')

## Statistical Data 

master_neonates_ids = pd.read_csv("neonates_admissions.csv")

master_neonates_ids = pd.merge(master_neonates_ids, master_neonates, on='HADM_ID', how='inner')

master_neonates_ids.columns 

master_neonates_ids.drop(['subject_id', 'DISCHTIME', 'ETHNICITY',
       'GENDER', 'EXPIRE_FLAG'], axis=1, inplace=True)

master_neonates_ids.head(2)

master_neonates_ids.shape 

## 4.1 DateTimeEvents 

datetimeevents_df = pd.read_csv("Data_3173_Stat/datetimeevents.csv")

datetimeevents_df.head()

datetimeevents_df = pd.merge(datetimeevents_df, master_neonates_ids, on='HADM_ID', how='inner')

datetimeevents_df.head(2)

datetimeevents_df.shape 

datetimeevents_df['CHARTTIME'] = datetimeevents_df['CHARTTIME'].astype('datetime64[ns]')
datetimeevents_df['ADMITTIME'] = datetimeevents_df['ADMITTIME'].astype('datetime64[ns]') 
    
## Calculating time interval 
datetimeevents_df['TIME_INTERVAL'] = datetimeevents_df['CHARTTIME'] - datetimeevents_df['ADMITTIME']

## Extracting only 24 hours data 
datetimeevents_24 = datetimeevents_df[(datetimeevents_df.TIME_INTERVAL <= '1 days') & (datetimeevents_df.TIME_INTERVAL >= '-1 days')]

datetimeevents_24.drop(['TIME_INTERVAL'], axis=1, inplace=True) 
    
## Saving into a given file 
datetimeevents_24.to_csv("Data_3145_Stat_24/datetimeevents_24.csv", index=False)

datetimeevents_24.shape

## 4.2 InputEvents 

inputevents_df = pd.read_csv("Data_3173_Stat/neonates_inputevents.csv")

inputevents_df.head(2)

inputevents_df.shape 

inputevents_df = pd.merge(inputevents_df, master_neonates_ids, on='HADM_ID', how='inner')

inputevents_df['CHARTTIME'] = inputevents_df['CHARTTIME'].astype('datetime64[ns]')
inputevents_df['ADMITTIME'] = inputevents_df['ADMITTIME'].astype('datetime64[ns]') 
    
## Calculating time interval 
inputevents_df['TIME_INTERVAL'] = inputevents_df['CHARTTIME'] - inputevents_df['ADMITTIME']

## Extracting only 24 hours data 
inputevents_24 = inputevents_df[(inputevents_df.TIME_INTERVAL <= '1 days') & (inputevents_df.TIME_INTERVAL >= '-1 days')]

inputevents_24.drop(['TIME_INTERVAL'], axis=1, inplace=True) 
    
## Saving into a given file 
inputevents_24.to_csv("Data_3145_Stat_24/inputevents_24.csv", index=False)

inputevents_24.shape 

inputevents_24['HADM_ID'].nunique()

## 4.3 Microbiologyevents

microbiologyevents_df = pd.read_csv("Data_3173_Stat/neonates_microbiologyevents.csv")
microbiologyevents_df.head(2)

microbiologyevents_df.shape 

microbiologyevents_df = pd.merge(microbiologyevents_df, master_neonates_ids, on='HADM_ID', how='inner')

microbiologyevents_df['CHARTTIME'] = microbiologyevents_df['CHARTTIME'].astype('datetime64[ns]')
microbiologyevents_df['ADMITTIME'] = microbiologyevents_df['ADMITTIME'].astype('datetime64[ns]') 
    
## Calculating time interval 
microbiologyevents_df['TIME_INTERVAL'] = microbiologyevents_df['CHARTTIME'] - microbiologyevents_df['ADMITTIME']

## Extracting only 24 hours data 
microbiologyevents_24 = microbiologyevents_df[(microbiologyevents_df.TIME_INTERVAL <= '1 days') & (microbiologyevents_df.TIME_INTERVAL >= '-1 days')]

microbiologyevents_24.drop(['TIME_INTERVAL'], axis=1, inplace=True) 
    
## Saving into a given file 
microbiologyevents_24.to_csv("Data_3145_Stat_24/microbiologyevents_24.csv", index=False)

microbiologyevents_24.shape

## 4.4 Outputevents

outputevents_df = pd.read_csv("Data_3173_Stat/neonates_outputevents.csv")
outputevents_df.head(2)

outputevents_df.shape 

outputevents_df = pd.merge(outputevents_df, master_neonates_ids, on='HADM_ID', how='inner')

outputevents_df['CHARTTIME'] = outputevents_df['CHARTTIME'].astype('datetime64[ns]')
outputevents_df['ADMITTIME'] = outputevents_df['ADMITTIME'].astype('datetime64[ns]') 
    
## Calculating time interval 
outputevents_df['TIME_INTERVAL'] = outputevents_df['CHARTTIME'] - outputevents_df['ADMITTIME']

## Extracting only 24 hours data 
outputevents_24 = outputevents_df[(outputevents_df.TIME_INTERVAL <= '1 days') & (outputevents_df.TIME_INTERVAL >= '-1 days')]

outputevents_24.drop(['TIME_INTERVAL'], axis=1, inplace=True) 
    
## Saving into a given file 
outputevents_24.to_csv("Data_3145_Stat_24/outputevents_24.csv", index=False)

outputevents_24.shape 

## 4.5 Prescriptions

prescriptions_df = pd.read_csv("Data_3173_Stat/neonates_prescriptions.csv")
prescriptions_df.head(2)

prescriptions_df = pd.merge(prescriptions_df, master_neonates_ids, on='HADM_ID', how='inner')

prescriptions_df['STARTDATE'] = prescriptions_df['STARTDATE'].astype('datetime64[ns]')
prescriptions_df['ADMITTIME'] = prescriptions_df['ADMITTIME'].astype('datetime64[ns]') 
    
## Calculating time interval 
prescriptions_df['TIME_INTERVAL'] = prescriptions_df['STARTDATE'] - prescriptions_df['ADMITTIME']

## Extracting only 24 hours data 
prescriptions_24 = prescriptions_df[(prescriptions_df.TIME_INTERVAL <= '1 days') & (prescriptions_df.TIME_INTERVAL >= '-1 days')]

prescriptions_24.drop(['TIME_INTERVAL'], axis=1, inplace=True) 
    
## Saving into a given file 
prescriptions_24.to_csv("Data_3145_Stat_24/prescriptions_24.csv", index=False)
