==> data_preprocessing/Step 8  - Single Value Data from Chartevents.ipynb <==
# Single Value Data from Chartevents
* by Firuz Juraev (PhD student)

import pandas as pd 
from datetime import datetime
import seaborn as sns
import glob
from os import listdir 
import os.path
from os import path
import numpy as np 
import warnings
warnings.filterwarnings('ignore')

def side_by_side(*objs, **kwds):
    from pandas.io.formats.printing import adjoin
    space = kwds.get('space', 4)
    reprs = [repr(obj).split('\n') for obj in objs]
    print (adjoin(space, *reprs))
    print()
    return

### Variables 

directory = "Data_3145_SV_24/" 

bad_neonates = [100044, 100062, 100081, 159338, 155826, 187822, 145130, 146110, 120546, 157469]

### Functions 

def list_files(directory):
    files = []
    counter = 0
    for f in listdir(directory):
        if f.endswith('.' + "csv"):
            files.append(f)
            counter = counter + 1
    print ("CSV Files: " + str(counter))
    return files

def drop_defected_neonates_rows(df): 
    for neonate in bad_neonates: 
        drop_list = []

        i = df[df['HADM_ID'] == neonate].index                             

        # Appending indeces to drop list  
        drop_list.append(i) 

        ## droping rows 
        for i in drop_list: 
            df.drop(i, axis=0, inplace=True)
        
    side_by_side(df.isnull().sum(), df.count()) 
        
    return df 

def merge_blood_test(neonates_df, df, test_type): 
    df = df.groupby("HADM_ID")

    mean_df = df.mean()
    mean_df = mean_df.reset_index()

    mean_df.drop(['subject_id'], axis=1, inplace=True)
    
    mean_df.rename(columns={'VALUENUM':test_type}, inplace=True) 
    
    # Merge neonates with blood test  
    neonates_df = pd.merge(neonates_df, mean_df, on='HADM_ID', how='inner')
    
    print (test_type + " is added!")
    print ("Unique Neonates: " + str(neonates_df['HADM_ID'].nunique()))
    side_by_side(neonates_df.isnull().sum(), neonates_df.count()) 
    
    return neonates_df

### Run 

list_files(directory)

###  1.1 Birth Weight 

birth_weight = pd.read_csv(directory + "birth_weight_24.csv")

birth_weight.head(2)

birth_weight.shape 

birth_weight = drop_defected_neonates_rows(birth_weight) 

birth_weight.describe()

birth_weight.head(2)

birth_weight.columns 

birth_weight.drop(['subject_id', 'CHARTTIME', 'ADMITTIME'], axis=1, inplace=True)

birth_weight.head()

birth_weight.shape 

#### Save to File 

birth_weight.to_csv("FinalData/SingleValue/birth_weight.csv", index=False)

### 1.2 Head Circ  

head_circ = pd.read_csv(directory + "head_circ_24.csv")

head_circ.head(2) 

head_circ = drop_defected_neonates_rows(head_circ) 

head_circ.describe()

head_circ.head(2)

head_circ.shape 

head_circ.drop(['subject_id', 'CHARTTIME', 'ADMITTIME'], axis=1, inplace=True)

head_circ.head(2)

#### Save to File 

head_circ.to_csv("FinalData/SingleValue/head_circ.csv", index=False)

## Blood Test 

* bands
* monos 
* eosinophils
* lymphs 
* neuts 
* platelet

neonates = (head_circ['HADM_ID']).to_frame()

neonates.shape 

### 2.1 Bands 

bands = pd.read_csv(directory + "bands_24.csv") 

bands.head(2)

bands = drop_defected_neonates_rows(bands) 

bands['HADM_ID'].nunique()

neonates = merge_blood_test(neonates, bands, "BANDS") 

### 2.2 MONOs

monos = pd.read_csv(directory + "monos_24.csv") 

monos.head(2)

monos = drop_defected_neonates_rows(monos) 

monos['HADM_ID'].nunique()

monos[monos['VALUENUM'].isnull() == True]

monos[monos['HADM_ID'] == 121852]

monos.dropna(inplace=True)

monos[monos['HADM_ID'] == 121852]

neonates = merge_blood_test(neonates, monos, "MONOs") 

### 2.3 Eosinophils

eosinophils = pd.read_csv(directory + "eosinophils_24.csv") 

eosinophils.head(2)

eosinophils = drop_defected_neonates_rows(eosinophils) 

eosinophils[eosinophils.VALUENUM.isnull() == True]

eosinophils[eosinophils.HADM_ID == 121852]

eosinophils.dropna(inplace=True)

neonates = merge_blood_test(neonates, eosinophils, "EOSINOPHILS") 

### 2.4 NEUTS 

neuts = pd.read_csv(directory + "neuts_24.csv") 

neuts.head(2)

neuts = drop_defected_neonates_rows(neuts) 

neuts[neuts.VALUENUM.isnull() == True]

neuts[neuts.HADM_ID == 121852]

neuts.dropna(inplace=True)

neonates = merge_blood_test(neonates, neuts, "NEUTS") 

### 2.5 Lymphs

lymphs = pd.read_csv(directory + "lymphs_24.csv") 

lymphs.head(2)

lymphs.info()

lymphs = drop_defected_neonates_rows(lymphs) 

lymphs['HADM_ID'].nunique()

lymphs[lymphs.VALUENUM.isnull() == True]

lymphs[lymphs.HADM_ID == 121852]

lymphs.drop(['Unnamed: 0', 'ROW_ID', 'VALUE', 'VALUEUOM'], axis=1, inplace=True)

lymphs.head(2)

lymphs.dropna(inplace=True)

lymphs[lymphs.VALUENUM.isnull() == True]

lymphs['HADM_ID'].nunique()

neonates = merge_blood_test(neonates, lymphs, "LYMPHS") 

### 2.6 PLATELET

platelet = pd.read_csv(directory + "platelet_24.csv") 

platelet.head(2)

platelet.drop(['Unnamed: 0'], axis=1, inplace=True)

platelet.head(2)

platelet = drop_defected_neonates_rows(platelet) 

platelet = drop_defected_neonates_rows(platelet) 

platelet['HADM_ID'].nunique()

platelet[platelet.VALUENUM.isnull() == True]

platelet[platelet.HADM_ID == 157469]

platelet.dropna(inplace=True)

platelet['HADM_ID'].nunique()

### Fix 

neonates = drop_defected_neonates_rows(neonates) 

neonates = merge_blood_test(neonates, platelet, "PLATELET") 

neonates.head()

neonates.shape 

neonates.to_csv("FinalData/SingleValue/blood_test_features.csv", index=False)
