==> data_preprocessing/Step 10 - Merging Single Value Features .ipynb <==
# Merging Single Value Features

* by Firuz Juraev (Combined Master/PhD student)

import pandas as pd 
from datetime import datetime
import seaborn as sns
import glob
from os import listdir 
import os.path
from os import path
import numpy as np 

from sklearn.preprocessing import OrdinalEncoder

import warnings
warnings.filterwarnings('ignore')

directory = "FinalData/SingleValue/" 

### Functions 

def side_by_side(*objs, **kwds):
    from pandas.io.formats.printing import adjoin
    space = kwds.get('space', 4)
    reprs = [repr(obj).split('\n') for obj in objs]
    print (adjoin(space, *reprs))
    print()
    return

def list_files(directory):
    files = []
    counter = 0
    for f in listdir(directory):
        if f.endswith('.' + "csv"):
            files.append(f)
            counter = counter + 1
    print ("CSV Files: " + str(counter))
    return files

list_files(directory)

### Load Patients' Data 

patients = pd.read_csv(directory + "patients_single_value_data.csv") 

patients.head(2)

### Birth Weight 

birth_weight = pd.read_csv(directory + "birth_weight.csv")  

birth_weight.head(2)

birth_weight.rename(columns = {'VALUENUM': 'BIRTH_WEIGHT'}, inplace=True) 

patients = pd.merge(patients, birth_weight, on='HADM_ID', how='inner') 

patients.head(2)

### Head Circ 

head_circ = pd.read_csv(directory + "head_circ.csv")  

head_circ.head(2)

head_circ.rename(columns = {'VALUENUM': 'HEAD_CIRC'}, inplace=True) 

patients = pd.merge(patients, head_circ, on='HADM_ID', how='inner') 

patients.head(2)

### Blood Test 

blood_test = pd.read_csv(directory + "blood_test_features.csv")  

blood_test.head(2)

patients = pd.merge(patients, blood_test, on='HADM_ID', how='inner') 

patients.head(2)

## Some filtering  

patients.drop(['DISCHARGE_LOCATION'], axis=1, inplace=True)

patients.rename(columns={'EXPIRE_FLAG': 'DEAD'}, inplace=True)

ord_encoder = OrdinalEncoder()

#convert categorical columns to int  

patients["ETHNICITY"] = ord_encoder.fit_transform(patients[["ETHNICITY"]])  

patients.head()

patients.columns

patients = patients[['SUBJECT_ID', 'HADM_ID', 'DOB', 'GENDER', 'ETHNICITY', 'ADMITTIME',
       'DISCHTIME', 'DATEEVENTS', 'DW10_COUNT', 'D10W_SUM',
       'D10W_MEAN', 'URINE_COUNT', 'URINE_SUM', 'URINE_AVG',
       'MICROBIOLOGY_TEST', 'NEGATIVE_RESULT', 'PRESCRIPTIONS',
       'BIRTH_WEIGHT', 'HEAD_CIRC', 'BANDS', 'MONOs', 'EOSINOPHILS', 'NEUTS',
       'LYMPHS', 'PLATELET', 'DEAD', 'POD', 'LOS']]
                    
patients.head()

patients.to_csv("FinalData/NeonatesSingleValues_FOR_STATISTICS.csv", index=False) 

patients.drop(['ADMITTIME', 'DOB', 'DISCHTIME'], axis=1, inplace=True)

patients.head(2)

patients.rename(columns={'SUBJECT_ID': 'subject_id'}, inplace=True)

### Submit 

patients.to_csv("FinalData/NeonatesSingleValues.csv", index=False)

THE END
