==> data_preprocessing/Step 11 - Merge Time Series .ipynb <==
# Merge Time Series Features 

* by Firuz Juraev (Combined Master/PhD student)

import pandas as pd 
from datetime import datetime
import seaborn as sns
import glob
from os import listdir 
import os.path
from os import path
import numpy as np 
import warnings
warnings.filterwarnings('ignore')

def side_by_side(*objs, **kwds):
    from pandas.io.formats.printing import adjoin
    space = kwds.get('space', 4)
    reprs = [repr(obj).split('\n') for obj in objs]
    print (adjoin(space, *reprs))
    print()
    return

def list_files(directory):
    files = []
    counter = 0
    for f in listdir(directory):
        if f.endswith('.' + "csv"):
            files.append(f)
            counter = counter + 1
    print ("CSV Files: " + str(counter))
    return files

### Define variables 

directory_FB = "FinalData/TimeSeries/Clean_Data_FB/" 
directory_Mean = "FinalData/TimeSeries/Clean_Data_Mean/"
directory_NaN = "FinalData/TimeSeries/Clean_Data_NaN/"

Main_directory = "FinalData/"

### Functions 

def mergeTS(df_list, df_name_list, filename, los_df):
    df_list[0].rename(columns={'VALUENUM':df_name_list[0]}, inplace=True)
    
    for i in range(1, len(df_list)): 
        df_list[i].drop(['HADM_ID', 'subject_id', 'CHARTTIME'], axis=1, inplace=True) 
        df_list[i].rename(columns={'VALUENUM':df_name_list[i]}, inplace=True)
    
    df = df_list[0]
    
    for i in range(1, len(df_list)): 
        ## Concatining 
        df = pd.concat([df, df_list[i]], axis=1)
    
    df['CHARTTIME'] = df.groupby(['subject_id'])['CHARTTIME'].cumcount() 
    
    side_by_side(df.isnull().sum(), df.count())
    
    df = pd.merge(df, los_df, on='HADM_ID', how='inner')
    
    if path.exists(filename): 
        df.to_csv(filename, mode='a', header=False, index=False)
    else: 
        df.to_csv(filename, index=False, header=True)
    
    print ("200OK - Success")

def merging_data_in_folder(directory, los_df, filename): 
    df_list = []
    df_name_list = [] 
    
    files = list_files(directory) 
     
    for file in files: 
        df_name_list.append(file.replace(".csv", ""))  
        
        df = pd.read_csv(directory + file) 
        df_list.append(df) 
    
    mergeTS(df_list, df_name_list, Main_directory + filename, los_df)

### Run 

neonates = pd.read_csv("FinalData/NeonatesSingleValues.csv") 

neonates.head(2)

neonates.columns

neonates.drop(['subject_id', 'GENDER', 'ETHNICITY', 'DATEEVENTS',
       'DW10_COUNT', 'D10W_SUM', 'D10W_MEAN', 'URINE_COUNT', 'URINE_SUM',
       'URINE_AVG', 'MICROBIOLOGY_TEST', 'NEGATIVE_RESULT', 'PRESCRIPTIONS',
       'BIRTH_WEIGHT', 'HEAD_CIRC', 'BANDS', 'MONOs', 'EOSINOPHILS', 'NEUTS',
       'LYMPHS', 'PLATELET', 'DEAD', 'POD'], axis=1, inplace=True)

=======================================================================================

## Merging Time Series Features [Run]

neonates.head(1)

filename_FB   = "NeonatesTimeSeries_FB.csv"
filename_Mean = "NeonatesTimeSeries_Mean.csv"  
filename_NaN  = "NeonatesTimeSeries_NaN.csv" 

merging_data_in_folder(directory_FB, neonates, filename_FB)
merging_data_in_folder(directory_Mean, neonates, filename_Mean)
merging_data_in_folder(directory_NaN, neonates, filename_NaN)

df = pd.read_csv(Main_directory + filename_FB) 

df.head()
