==> data_preprocessing/Step 12 - Convert Time Series into Single.ipynb <==
## Converting Time Series Data into Single Value Data 

- Firuz Juraev 

import pandas as pd 
from os import path
import numpy as np 
import math 

def side_by_side(*objs, **kwds):
    from pandas.io.formats.printing import adjoin
    space = kwds.get('space', 4)
    reprs = [repr(obj).split('\n') for obj in objs]
    print (adjoin(space, *reprs))
    print()
    return

data = pd.read_csv('NeonatesTimeSeries_NaN.csv')

data.head()

## Converting Function

def compress(column_name, subject_id, column_values):
    if not math.isnan(column_values.std()): 
        std_value = column_values.std()
    else: 
        std_value = 0
    mean_value = column_values.mean()
    median_value = column_values.median()
    if len(column_values.mode().values) > 0: 
        mode_value = column_values.mode().values[0]
    else: 
        mode_value = np.nan 
     
    
    # create columns     
    slice_columns = ["subject_id", 
               (column_name + "_mean"),
               (column_name + "_median"),
               (column_name + "_mode"), 
               (column_name + "_std")]
    
    # Creating a row 
    df = pd.DataFrame(columns = slice_columns)
    df.loc[0] = [subject_id, mean_value, median_value, mode_value, std_value]
    
    return df 

def converter(df, filename): 
    df_columns = data.columns.to_list()[3:-1]
    
    parts_list = [] 
    
    for col in df_columns: 
        part = compress(col, df['subject_id'].iloc[0], df[col])
        parts_list.append(part) 
    
    join_dfs = parts_list[0] 
    
    for i in range(1, len(parts_list)): 
        join_dfs = pd.merge(join_dfs, parts_list[i], on='subject_id', how='inner') 
    
    if path.exists(filename): 
        join_dfs.to_csv(filename, mode='a', header=False, index=False)
    else: 
        join_dfs.to_csv(filename, index=False, header=True)
              

df_columns.to_list()[3:-1] 

# finding unique ids 
unique_subjects_ids = data['subject_id'].unique() 

# Converting 
for i in range(0, len(unique_subjects_ids)): 
    converter(data.loc[data['subject_id'] == unique_subjects_ids[i]], 'test3.csv')         

test = pd.read_csv('test3.csv')
test.head(10)

side_by_side(test.isnull().sum(), test.count())
